+++ tests/Hmac/HmacTokenCsrfMiddlewareTest.php
@@ -6,6 +6,7 @@ namespace Yiisoft\Csrf\Tests\Hmac;
 
 use Yiisoft\Csrf\CsrfTokenInterface;
 use Yiisoft\Csrf\Hmac\HmacCsrfToken;
+use Yiisoft\Csrf\Tests\Hmac\IdentityGenerator\MockCsrfTokenIdentityGenerator;
 use Yiisoft\Csrf\Tests\TokenCsrfMiddlewareTest;
 use Yiisoft\Security\Random;
