+++ tests/Synchronizer/Generator/RandomCsrfTokenGeneratorTest.php
@@ -2,10 +2,10 @@
 
 declare(strict_types=1);
 
-namespace Yiisoft\Csrf\Tests\Synchronizer;
+namespace Yiisoft\Csrf\Tests\Synchronizer\Generator;
 
 use PHPUnit\Framework\TestCase;
-use Yiisoft\Csrf\Synchronizer\RandomCsrfTokenGenerator;
+use Yiisoft\Csrf\Synchronizer\Generator\RandomCsrfTokenGenerator;
 
 final class RandomCsrfTokenGeneratorTest extends TestCase
 {
